+++ src/main/java/com/elmakers/mine/bukkit/wand/WandAction.java
@@ -1,0 +1,5 @@
+package com.elmakers.mine.bukkit.wand;
+
+public enum WandAction {
+    NONE, CAST, TOGGLE, CYCLE, CYCLE_REVERSE, CYCLE_HOTBAR, CYCLE_HOTBAR_REVERSE 
+}